requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.5.2
cx-Oracle==6.4.1
feedparser==5.2.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
metar==1.6.0
python-decouple==3.1
PyYAML==3.13
requests==2.19.1
six==1.11.0
swagger-spec-validator==2.4.0
urllib3==1.23
waitress==1.1.0
Werkzeug==0.14.1
